Workaround Things not supporting historic completed to-dos.
When importing a to-do with completed: true and when set to a time in
the past, Things will set the completion time of the new item to be the
current time.

While this is the case we can try to work around it by updating the
Things sqlite3 database directly after the items have been created.
The TMTask.stopDate appears to record the completion time. This change
builds a SQL script to update imported items completion times.