Use JXA instead of SQL to update completion times
The SQL approach didn't update the timestamps needed for Things to
propagate changes to Things Cloud, and solving that feels like
digging a deeper hole. At the same time, using the JavaScript for
Automation scripting interface into Things allows the completion
time to be updated into the past, and is (presumably) supported.

All told, it might actually make more sense to generate a JXA
script to create the projects and to-dos in the first place,
instead of going through the URL scheme.